Refactor handling of changes
