+++ indico/modules/events/timetable/client/js/reducers.js
@@ -18,7 +18,7 @@ import {
   changeSessionColor,
   changeBreakColor,
   dropUnscheduledContribs,
-} from './util';
+} from './operations';
 
 const entryTypeMapping = {
   Session: 'session',
@@ -56,7 +56,7 @@ const preprocessData = (data, eventInfo) => {
       type: entryTypeMapping[e.entryType],
       start: e.startDate && new Date(Date.parse(`${e.startDate.date} ${e.startDate.time}`)),
       attachmentCount: e.attachments?.files?.length + e.attachments?.folders.length, // TODO count files in folders
-      color: e.entryType === 'Break' ? {text: e.textColor, background: e.color} : undefined,
+      color: e.entryType === 'Break' ? {text: e.textColor, background: e.color} : null,
       deleted: false,
     }))
   );
@@ -90,7 +90,7 @@ export default {
   ) => {
     switch (action.type) {
       case actions.SET_TIMETABLE_DATA:
-        return {...state, ...preprocessEntries(...preprocessData(action.data, action.eventInfo))};
+        return preprocessEntries(state, ...preprocessData(action.data, action.eventInfo));
       case actions.MOVE_ENTRY:
         return {...state, ...moveEntry(state, action.args)};
       case actions.RESIZE_ENTRY:
